+++ R/groupby.R
@@ -33,7 +33,6 @@ GroupBy <- new.env(parent = emptyenv())
 #'
 #' @examples pl$DataFrame(iris)$groupby("Species")
 print.GroupBy = function(x, ...) {
-  cat("polars GroupBy: ")
   .pr$DataFrame$print(x)
   cat("groups: ")
   .pr$ProtoExprArray$print(attr(x,"private")$groupby_input)
@@ -270,11 +269,15 @@ GroupBy_null_count <- function() {
 #' @return R data.frame
 #' @export
 #'
-#' @examples pl$DataFrame(iris)$as_data_frame() #R-polars back and forth
-GroupBy_as_data_frame = function(...) {
+#' @examples pl$DataFrame(iris)$to_data_frame() #R-polars back and forth
+GroupBy_to_data_frame = function(...) {
   as.data.frame(
     x = unwrap(.pr$DataFrame$to_list(self)),
     col.names = .pr$DataFrame$columns(self),
     ...
   )
 }
+
+#' Alias to GroupBy_to_data_frame (backward compatibility)
+#' @noRd
+GroupBy_as_data_frame = GroupBy_to_data_frame