+++ tests/testthat/test-dataframe.R
@@ -56,24 +56,25 @@ expected_iris_select_df <- structure(list(miah = c(
 ))
 
 
+patrick::with_parameters_test_that("DataFrame, mixed input, create and print",
+  {
+    input_vectors_and_series <- list(
+      newname = pl$Series(c(1, 2, 3, 4, 5), name = "b"), # overwrite name b with newname
+      pl$Series((1:5) * 5, "a"),
+      pl$Series(letters[1:5], "b"),
+      c(5, 4, 3, 2, 1), # unnamed vector
+      named_vector = c(15, 14, 13, 12, 11), # named provide
+      c(5, 4, 3, 2, 0)
+    )
 
-
-
-
-test_that("DataFrame, mixed input, create and print", {
-  input_vectors_and_series <- list(
-    newname = pl$Series(c(1, 2, 3, 4, 5), name = "b"), # overwrite name b with newname
-    pl$Series((1:5) * 5, "a"),
-    pl$Series(letters[1:5], "b"),
-    c(5, 4, 3, 2, 1), # unnamed vector
-    named_vector = c(15, 14, 13, 12, 11), # named provide
-    c(5, 4, 3, 2, 0)
-  )
-
-  # clone into DataFrame and change one name
-  df <- pl$DataFrame(input_vectors_and_series)
-  expect_snapshot(df)
-})
+    # clone into DataFrame and change one name
+    df <- pl$DataFrame(input_vectors_and_series)
+    .env_var <- .value
+    names(.env_var) <- .name
+    withr::with_envvar(.env_var, expect_snapshot(df))
+  },
+  .cases = make_print_cases()
+)
 
 test_that("DataFrame, input free vectors, input empty", {
   # passing vector directly is equal to passing one
@@ -148,7 +149,7 @@ test_that("DataFrame, select sum over", {
   )
 
   expect_equal(
-    df$as_data_frame(),
+    df$to_data_frame(),
     expected_iris_select_df
   )
 
@@ -165,7 +166,7 @@ test_that("map unity", {
       pl$col("Sepal.Length")$
         map(\(s) s)
     )$
-    as_data_frame()[, 1, drop = FALSE]
+    to_data_frame()[, 1, drop = FALSE]
 
   # float is preserved
   expect_identical(
@@ -182,7 +183,7 @@ test_that("map unity", {
       pl$col("Sepal.Length")$
         map(\(s) s)
     )$
-    as_data_frame()[, 1, drop = FALSE]
+    to_data_frame()[, 1, drop = FALSE]
 
   expect_identical(
     x,
@@ -196,7 +197,7 @@ test_that("map unity", {
     pl$col("Species")$
       map(\(s) s)
   )$
-    as_data_frame()[, 1]
+    to_data_frame()[, 1]
 
   expect_different(x, iris[, 1, drop = FALSE])
 
@@ -207,7 +208,7 @@ test_that("map unity", {
     pl$col("Species")$
       map(\(s) s)
   )$
-    as_data_frame()
+    to_data_frame()
   expect_identical(x, iris[, 5, drop = FALSE])
 })
 
@@ -227,7 +228,7 @@ test_that("map type", {
       map(\(s) s * 25L)$
       map(\(s) s / 4)
   )$
-    as_data_frame()[, 1, drop = FALSE]
+    to_data_frame()[, 1, drop = FALSE]
 
   expect_identical(x, int_iris[, 1, drop = FALSE] * 25L / 4L)
 })
@@ -243,7 +244,7 @@ test_that("cloning", {
   # deep copy clone rust side object, hence not same mem address
   # For some reason, expect_identical(pf, pf3) fails
   pf3 <- pf$clone()
-  expect_identical(pf$as_data_frame(), pf3$as_data_frame())
+  expect_identical(pf$to_data_frame(), pf3$to_data_frame())
   expect_different(pl$mem_address(pf), pl$mem_address(pf3))
 })
 
@@ -268,7 +269,7 @@ test_that("get column(s)", {
   # }
 
 
-  # list_of_vectors = lapply(actual_list_of_series, function(x) x$to_r_vector())
+  # list_of_vectors = lapply(actual_list_of_series, function(x) x$to_vector())
   # expect_identical(
   #   list_of_vectors,
   #   as.list(iris)
@@ -319,12 +320,12 @@ test_that("with_columns lazy/eager", {
   rdf$`not c` <- !rdf$c
 
   expect_identical(
-    df_actual$as_data_frame(check.names = FALSE),
+    df_actual$to_data_frame(check.names = FALSE),
     rdf
   )
 
   expect_identical(
-    ldf_actual$collect()$as_data_frame(check.names = FALSE),
+    ldf_actual$collect()$to_data_frame(check.names = FALSE),
     rdf
   )
 
@@ -338,7 +339,7 @@ test_that("with_columns lazy/eager", {
   pl$reset_polars_options()
 
   expect_identical(
-    ldf_actual_kwarg_named$collect()$as_data_frame(check.names = FALSE),
+    ldf_actual_kwarg_named$collect()$to_data_frame(check.names = FALSE),
     rdf
   )
 })
@@ -352,34 +353,34 @@ test_that("limit lazy/eager", {
   )
   df <- pl$DataFrame(l)
   ldf <- df$lazy()
-  rdf <- df$as_data_frame()
+  rdf <- df$to_data_frame()
 
   expect_identical(
-    df$limit(2)$as_data_frame(),
+    df$limit(2)$to_data_frame(),
     rdf[1:2, ]
   )
 
   expect_identical(
-    ldf$limit(2)$collect()$as_data_frame(),
+    ldf$limit(2)$collect()$to_data_frame(),
     rdf[1:2, ]
   )
 
   # lazy bounds
-  expect_identical(df$limit(0)$as_data_frame(), rdf[integer(), ])
+  expect_identical(df$limit(0)$to_data_frame(), rdf[integer(), ])
   expect_error(ldf$limit(-1))
   expect_error(ldf$limit(2^32))
-  expect_identical(ldf$limit(2^32 - 1)$collect()$as_data_frame(), rdf)
+  expect_identical(ldf$limit(2^32 - 1)$collect()$to_data_frame(), rdf)
 
   # eager bounds
-  expect_identical(ldf$limit(0)$collect()$as_data_frame(), rdf[integer(), ])
+  expect_identical(ldf$limit(0)$collect()$to_data_frame(), rdf[integer(), ])
   expect_error(df$limit(-1))
   expect_error(df$limit(2^32))
-  expect_identical(df$limit(2^32 - 1)$as_data_frame(), rdf)
+  expect_identical(df$limit(2^32 - 1)$to_data_frame(), rdf)
 })
 
 
 
-test_that("to_Struct, unnest, to_frame, as_data_frame", {
+test_that("to_Struct, unnest, to_frame, to_data_frame", {
   # round-trip conversion from DataFrame with two columns
   df <- pl$DataFrame(
     a = 1:5,
@@ -396,10 +397,10 @@ test_that("to_Struct, unnest, to_frame, as_data_frame", {
   expect_identical(df$to_list(), df_s$to_list(unnest_structs = TRUE)[[1L]])
 
   # tedious way to unnest a data.frame of one column struct
-  df_e <- as.data.frame(do.call(rbind, df_s$as_data_frame()[[1L]])) |>
+  df_e <- as.data.frame(do.call(rbind, df_s$to_data_frame()[[1L]])) |>
     lapply(unlist) |>
     as.data.frame()
-  expect_identical(df$as_data_frame(), df_e)
+  expect_identical(df$to_data_frame(), df_e)
 })
 
 make_cases <- function() {
@@ -420,7 +421,7 @@ make_cases <- function() {
 
 patrick::with_parameters_test_that(
   "simple translations: eager", {
-    a = pl$DataFrame(mtcars)[[pola]]()$as_data_frame()
+    a = pl$DataFrame(mtcars)[[pola]]()$to_data_frame()
     b = data.frame(lapply(mtcars, base))
     testthat::expect_equal(a, b, ignore_attr = TRUE)
   },
@@ -428,23 +429,23 @@ patrick::with_parameters_test_that(
 )
 
 test_that("simple translations", {
-  a = pl$DataFrame(mtcars)$var(10)$as_data_frame()
+  a = pl$DataFrame(mtcars)$var(10)$to_data_frame()
   b = data.frame(lapply(mtcars, var))
   expect_true(all(a != b))
 
-  a = pl$DataFrame(mtcars)$std(10)$as_data_frame()
+  a = pl$DataFrame(mtcars)$std(10)$to_data_frame()
   b = data.frame(lapply(mtcars, sd))
   expect_true(all(a != b))
 
-  a = pl$DataFrame(mtcars)$reverse()$as_data_frame()
+  a = pl$DataFrame(mtcars)$reverse()$to_data_frame()
   b = mtcars[32:1,]
   expect_equal(a, b, ignore_attr = TRUE)
 
-  a = pl$DataFrame(mtcars)$slice(2, 4)$as_data_frame()
+  a = pl$DataFrame(mtcars)$slice(2, 4)$to_data_frame()
   b = mtcars[3:6,]
   expect_equal(a, b, ignore_attr = TRUE)
 
-  a = pl$DataFrame(mtcars)$slice(30)$as_data_frame()
+  a = pl$DataFrame(mtcars)$slice(30)$to_data_frame()
   b = tail(mtcars, 2)
   expect_equal(a, b, ignore_attr = TRUE)
 
@@ -466,12 +467,12 @@ test_that("null_count 64bit", {
   tmp = mtcars
   tmp[1:2, 1:2] = NA
   tmp[5, 3] = NA
-  a = pl$DataFrame(tmp)$null_count()$as_data_frame()
+  a = pl$DataFrame(tmp)$null_count()$to_data_frame()
   a = sapply(a, as.integer)
   b = sapply(tmp, function(x) sum(is.na(x)))
   expect_equal(a, b)
 
-  a = pl$DataFrame(tmp)$groupby("vs")$null_count()$as_data_frame()
+  a = pl$DataFrame(tmp)$groupby("vs")$null_count()$to_data_frame()
   expect_equal(dim(a), c(2, 11))
 })
 
@@ -480,7 +481,6 @@ test_that("tail", {
   b <- tail(mtcars)
   expect_equal(a, b, ignore_attr = TRUE)
 })
-
 
 test_that("drop_in_place", {
   dat = pl$DataFrame(iris)
@@ -580,5 +580,13 @@ test_that("unique", {
   expect_equal(x, 5)
   expect_equal(y, 5)
   expect_equal(z, 5)
+})
+
+test_that("as_data_frame (backward compatibility)", {
+  w <- as.data.frame(pl$DataFrame(mtcars)$to_data_frame())
+  x <- as.data.frame(pl$DataFrame(mtcars)$as_data_frame())
+  y <- mtcars
+  expect_equal(w, x, ignore_attr = TRUE)
+  expect_equal(w, y, ignore_attr = TRUE)
 })